Move token revocation methods to a trait
